+++ src/main/java/frc/robot/subsystems/swerve/module/stuffs/EncoderStuff.java
@@ -1,7 +1,7 @@
 package frc.robot.subsystems.swerve.module.stuffs;
 
 import edu.wpi.first.math.geometry.Rotation2d;
-import frc.robot.hardware.angleencoder.IAngleEncoder;
-import frc.robot.hardware.signal.InputSignal;
+import frc.robot.hard.interfaces.IAngleEncoder;
+import frc.robot.hard.interfaces.InputSignal;
 
 public record EncoderStuff(IAngleEncoder encoder, InputSignal<Rotation2d> positionSignal) {}
