+++ src/main/java/frc/robot/subsystems/swerve/module/stuffs/SteerStuff.java
@@ -1,9 +1,9 @@
 package frc.robot.subsystems.swerve.module.stuffs;
 
 import edu.wpi.first.math.geometry.Rotation2d;
-import frc.robot.hardware.motor.ControllableMotor;
-import frc.robot.hardware.request.IRequest;
-import frc.robot.hardware.signal.InputSignal;
+import frc.robot.hard.interfaces.ControllableMotor;
+import frc.robot.hard.interfaces.IRequest;
+import frc.robot.hard.interfaces.InputSignal;
 
 public record SteerStuff(
 	ControllableMotor steer,
